requirements.txt
django==3.2
djangorestframework==3.12.4
pymysql==1.0.2